Move optimalWidth and global to different files.
